refactor(utilities): move formatNumber/formatTime into @ripl/utilities
The formatters are zero-dependency Intl wrappers that only use utilities type
guards, so they belong in `@ripl/utilities`: `formatNumber` joins `number.ts`
and `formatTime` lands in a new `time.ts`. Remove the core `format` module and
its `@ripl/core` re-export, repoint the colour-legend consumer, and relocate the
tests.
