--- packages/core/src/format/index.ts
@@ -1,2 +1,0 @@
-export * from './number';
-export * from './time';
